updata compile locally

+++ main.py
@@ -1,7 +1,7 @@
-import unzip
 import os
 import sys
-import comparePDF, compileLocal
+import argparse
+import comparePDF, compileLocal, unzip, clear
 from wasmer import engine, Store, Module, Instance
 
 # Get the list of the zip files needed
@@ -19,21 +19,29 @@ def list_directories(directory):
             dir_list.append(os.path.join(root, dir))
     return dir_list
 
+# parameters
+parser = argparse.ArgumentParser(description="Process some files.")
+parser.add_argument('--mode', type=str, default='non-stop', help='Mode of operation')
+parser.add_argument('--engine', type=str, default='xelatex', help='LaTeX engine to use')
+args = parser.parse_args()
 # make sure we are in the right directory
 os.system("cd /Users/zhengbowen/nondefault/Intern/Arxtext")
-engine = "xelatex"
+
 ## define the directory and the unzip directory
+mode = args.mode
+engine = args.engine
 directory = "./test-files"
 unzip_dir = "./tem-unzipped-files"
 
 # unzip the files
-file_list = list_files(directory)
-for file in file_list:
-    if file.endswith(".zip"):
-        unzip.unzip_files(file, unzip_dir)
-    else:
-        print(f"Skipping {file}")
-os.system(f"rm -rf ./tem-unzipped-files/__MACOSX")
+if not os.path.exists(unzip_dir):
+    file_list = list_files(directory)
+    for file in file_list:
+        if file.endswith(".zip"):
+            unzip.unzip_files(file, unzip_dir)
+        else:
+            print(f"Skipping {file}")
+    os.system(f"rm -rf ./tem-unzipped-files/__MACOSX")
 
 dirs = list_directories(unzip_dir)
 # compile the files locally
@@ -43,10 +51,6 @@ for dir in dirs:
 # compare the files
 file_path1 = "./file1.pdf"
 file_path2 = "./file2.pdf"
-
-mode = "non-stop"
-if len(sys.argv) > 1:
-    mode = sys.argv[1]
 
 # if mode == "non-stop":
     # go through the whole directory and compare all files
@@ -60,4 +64,4 @@ else:
     print("The two PDFs are different.")
 
 # remove the unzipped files
-# os.system(f"rm -rf ./tem-unzipped-files")
+# clear.clear()
